src/main/java/easy/q20ValidParentheses/Solution.java
package easy.q20ValidParentheses;

import java.util.Map;
import java.util.Stack;

public class Solution {

    private final Map<Character, Character> map = Map.of(
            '(', ')',
            '[', ']',
            '{', '}'
    );

    public boolean isValid(String s) {
        Stack<Character> stack = new Stack<>();
        for (char ch : s.toCharArray()) {
            if (map.containsKey(ch)) {
                stack.push(map.get(ch));
            } else if (stack.isEmpty() || stack.pop() != ch) {
                return false;
            }
        }
        return stack.isEmpty();
    }
}
